refactor: split genhcl tests (#526)
